refactor(STADTPULS-385): move PublicSensorType
Back to hook that provides this data
